RIB-39 - Move UiEvent and UIActionEvents to different files
