+++ app/src/main/java/io/rafaelribeiro/spendless/presentation/screens/login/LoginActionEvent.kt
@@ -1,0 +1,5 @@
+package io.rafaelribeiro.spendless.presentation.screens.login
+
+sealed interface LoginActionEvent {
+    data class LoginSucceed(val username: String) : LoginActionEvent
+}
