+++ app/src/main/java/io/rafaelribeiro/spendless/presentation/screens/settings/SettingsActionEvent.kt
@@ -1,0 +1,9 @@
+package io.rafaelribeiro.spendless.presentation.screens.settings
+
+sealed interface SettingsActionEvent {
+    data object OnPreferencesClicked: SettingsActionEvent
+    data object OnSecurityClicked: SettingsActionEvent
+    data object OnLogoutClicked: SettingsActionEvent
+    data object OnAccountClicked: SettingsActionEvent
+    data object OnBackClicked: SettingsActionEvent
+}
