+++ app/src/main/java/io/rafaelribeiro/spendless/presentation/screens/settings/SettingsViewModel.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import io.rafaelribeiro.spendless.R
-import io.rafaelribeiro.spendless.core.data.TransactionCreator
 import io.rafaelribeiro.spendless.core.presentation.UiText
+import io.rafaelribeiro.spendless.domain.AuthRepository
 import io.rafaelribeiro.spendless.domain.UserSessionRepository
 import io.rafaelribeiro.spendless.domain.transaction.TransactionRepository
 import io.rafaelribeiro.spendless.domain.user.UserPreferencesRepository
@@ -16,6 +16,7 @@ import io.rafaelribeiro.spendless.presentation.screens.settings.SettingsActionEv
 import io.rafaelribeiro.spendless.presentation.screens.settings.SettingsActionEvent.OnPreferencesClicked
 import io.rafaelribeiro.spendless.presentation.screens.settings.SettingsActionEvent.OnSecurityClicked
 import io.rafaelribeiro.spendless.presentation.screens.settings.SettingsActionEvent.OnAccountClicked
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -24,13 +25,15 @@ import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 @HiltViewModel
 class SettingsViewModel @Inject constructor(
     private val userPreferencesRepository: UserPreferencesRepository,
     private val transactionRepository: TransactionRepository,
-    private val userSessionRepository: UserSessionRepository
+    private val userSessionRepository: UserSessionRepository,
+    private val authRepository: AuthRepository
 ) : ViewModel() {
 
     private val _uiState = MutableStateFlow(SettingsUiState())
@@ -52,8 +55,7 @@ class SettingsViewModel @Inject constructor(
             is SettingsUiEvent.AccountClicked -> sendActionEvent(OnAccountClicked)
             is SettingsUiEvent.BackClicked -> sendActionEvent(OnBackClicked)
             is SettingsUiEvent.DeleteAccountClicked -> deleteAccount()
-            is SettingsUiEvent.AddFakeTransactionsClicked -> addFakeTransactions()
-            is SettingsUiEvent.DeleteFakeTransactionsClicked -> deleteAllTransactions()
+            is SettingsUiEvent.DeleteTransactionsClicked -> deleteAllTransactions()
         }
     }
 
@@ -69,24 +71,21 @@ class SettingsViewModel @Inject constructor(
 
     private fun deleteAccount() {
         viewModelScope.launch {
-            transactionRepository.deleteAllTransactions()
-            userPreferencesRepository.clearAllPreferences()
+            withContext(Dispatchers.Default) {
+                transactionRepository.deleteAllTransactions()
+                userPreferencesRepository.clearAllPreferences()
+                authRepository.deleteAccount()
+            }
             userSessionRepository.updateSessionState(UserSessionState.Idle)
         }
     }
 
     private fun deleteAllTransactions() {
         viewModelScope.launch {
-            transactionRepository.deleteAllTransactions()
+            withContext(Dispatchers.Default) {
+                transactionRepository.deleteAllTransactions()
+            }
             showMessage(UiText.StringResource(R.string.transactions_deleted))
-        }
-    }
-
-    private fun addFakeTransactions() {
-        viewModelScope.launch {
-            val transactions = TransactionCreator.createTransactions(quantity = 100)
-            transactions.forEach { transactionRepository.saveTransaction(it) }
-            showMessage(UiText.StringResource(R.string.fake_transactions_created))
         }
     }
 
@@ -97,24 +96,4 @@ class SettingsViewModel @Inject constructor(
             updateUiState { it.copy(message = UiText.Empty, isError = false) }
         }
     }
-}
-
-sealed interface SettingsActionEvent {
-    data object OnPreferencesClicked: SettingsActionEvent
-    data object OnSecurityClicked: SettingsActionEvent
-    data object OnLogoutClicked: SettingsActionEvent
-    data object OnAccountClicked: SettingsActionEvent
-    data object OnBackClicked: SettingsActionEvent
-}
-
-sealed interface SettingsUiEvent {
-    data object PreferencesClicked: SettingsUiEvent
-    data object SecurityClicked: SettingsUiEvent
-    data object LogoutClicked: SettingsUiEvent
-    data object AccountClicked: SettingsUiEvent
-    data object BackClicked: SettingsUiEvent
-
-    data object DeleteAccountClicked: SettingsUiEvent
-    data object AddFakeTransactionsClicked: SettingsUiEvent
-    data object DeleteFakeTransactionsClicked: SettingsUiEvent
 }
