+++ app/src/main/java/io/rafaelribeiro/spendless/presentation/screens/settings/preferences/SettingsPreferencesActionEvent.kt
@@ -1,0 +1,5 @@
+package io.rafaelribeiro.spendless.presentation.screens.settings.preferences
+
+sealed interface SettingsPreferencesActionEvent {
+    data object OnBackClicked : SettingsPreferencesActionEvent
+}
