+++ app/src/main/java/io/rafaelribeiro/spendless/presentation/screens/settings/security/SettingsSecurityActionEvent.kt
@@ -1,0 +1,6 @@
+package io.rafaelribeiro.spendless.presentation.screens.settings.security
+
+sealed interface SettingsSecurityActionEvent {
+    data object OnBackClicked : SettingsSecurityActionEvent
+    data object RestartSession : SettingsSecurityActionEvent
+}
