+++ app/src/main/java/io/rafaelribeiro/spendless/presentation/screens/transactions/TransactionsUiEvent.kt
@@ -1,0 +1,7 @@
+package io.rafaelribeiro.spendless.presentation.screens.transactions
+
+sealed interface TransactionsUiEvent {
+    data class TransactionNoteClicked(val transactionId: Long) : TransactionsUiEvent
+    data object AddTransactionClicked : TransactionsUiEvent
+    data object DownloadTransactionsClicked : TransactionsUiEvent
+}
